+++ examples/demo/src/App.vue
@@ -84,6 +84,7 @@ const flags = ref({
   showNotToggle: true,
   showShiftActions: true,
   showCloneButtons: true,
+  showUngroupButtons: true,
   showLockButtons: true,
   showMuteButtons: true,
   showUndoRedo: true,